requirements.txt
linkify-it-py==2.0.2
markdown-it-py==3.0.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
Pygments==2.17.2
python-nmap==0.7.1
typing_extensions==4.9.0
uc-micro-py==1.0.2
